Add files via upload
A guided Codecademy project practicing working with Python strings. In the project, the student is corresponding with "Vishal," a more experienced coder who has an interest in ciphers. The student must develop Python code to decode (and encode) Caesar Ciphers knowing the offset, and then brute force a decode by trying all possible offsets. The second part of the project is developing a decoder function for Vigenere Ciphers. As you'll see in my code, I couldn't get the logic right, so I finally resorted to using the solution code. The final task was to develop a function to encode a Vigenere Cipher and send a message back to Vishal.